Solving E and Z-para in pardiso
